+++ apps/web/app/(feed)/bookmark/page.ts
@@ -1,1 +1,1 @@
-export { BookmarkPage as default } from '@/pages';
+export { BookmarkPage as default } from '@/pages/bookmark';
