+++ apps/web/app/(feed)/read-later/page.ts
@@ -1,1 +1,1 @@
-export { ReadLaterPage as default } from '@/pages';
+export { ReadLaterPage as default } from '@/pages/read-later';
