+++ apps/web/src/app/layouts/index.ts
@@ -1,0 +1,1 @@
+export { RootLayout } from './root-layout';
