--- apps/web/src/pages/bookmark/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as BookmarkPage } from './page';
