--- apps/web/src/pages/feed/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as FeedPage } from './page';
