+++ apps/web/src/pages/feed/ui/index.tsx
@@ -1,0 +1,3 @@
+export function FeedPage() {
+  return <div>FeedPage</div>;
+}
