--- apps/web/src/pages/feed/ui/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>FeedPage</div>;
-}
