--- apps/web/src/pages/history/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as HistoryPage } from './page';
