+++ apps/web/src/pages/history/ui/index.tsx
@@ -1,3 +1,3 @@
-export default function Page() {
+export function HistoryPage() {
   return <div>HistoryPage</div>;
 }
