--- apps/web/src/pages/index.ts
@@ -1,6 +1,0 @@
-export * from './bookmark';
-export * from './feed';
-export * from './history';
-export * from './main';
-export * from './read-later';
-export * from './tag';
