--- apps/web/src/pages/main/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as MainPage } from './page';
