+++ apps/web/src/pages/main/ui/index.tsx
@@ -1,0 +1,3 @@
+export function MainPage() {
+  return <div>MainPage</div>;
+}
