--- apps/web/src/pages/main/ui/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>MainPage</div>;
-}
