--- apps/web/src/pages/read-later/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ReadLaterPage } from './page';
