--- apps/web/src/pages/tag/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as TagPage } from './page';
