+++ apps/web/src/pages/tag/ui/index.tsx
@@ -1,0 +1,3 @@
+export function TagPage() {
+  return <div>TagPage</div>;
+}
