--- apps/web/src/pages/tag/ui/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>TagPage</div>;
-}
